+++ flag/service/github/github.go
@@ -1,0 +1,5 @@
+package github
+
+type GitHub struct {
+	Token string
+}